[update] securisation des pages, pour empecher d'y acceder via l'url, et attribution des roles pour la suppression d'article

+++ view/article/detail.php
@@ -14,14 +14,17 @@
             <img style="border-radius: 8%;" src="/upload/<?= htmlspecialchars($article['image_filename']) ?>" alt="Image de l'article">
             <p><?= nl2br(htmlspecialchars($article['textOfArticle'])) ?></p>
             <p>Posté le : <?= htmlspecialchars($article['date']) ?></p>
-            <a href="/ctrl/article/deleteDetails.php?id=<?= htmlspecialchars($article['id']) ?>" onclick="return confirm('Êtes-vous sûr de vouloir supprimer cet article ?');">
-                <button class="buttonDelete">Supprimer</button>
-            </a>
-            <a href="/ctrl/article/update-display.php?id=<?= htmlspecialchars($article['id']) ?>"><button class="buttonUpdate"><img class="iconeCorbeille" src="/asset/img/editer.png" alt=""></button></a>
+
+            <?php if (isset($_SESSION['user']) && ($_SESSION['user']['id'] == $article['idUser'] || $_SESSION['user']['idRole'] == '10')): ?>
+                <a href="/ctrl/article/deleteDetails.php?id=<?= htmlspecialchars($article['id']) ?>" onclick="return confirm('Êtes-vous sûr de vouloir supprimer cet article ?');">
+                    <button class="buttonDelete">Supprimer</button>
+                </a>
+                <a href="/ctrl/article/update-display.php?id=<?= htmlspecialchars($article['id']) ?>"><button class="buttonUpdate"><img class="iconeCorbeille" src="/asset/img/editer.png" alt=""></button></a>
+            <?php endif; ?>
         </article>
 
         <!-- Formulaire de commentaire -->
-         <?php if (isset($_SESSION['user'])): ?>
+        <?php if (isset($_SESSION['user'])): ?>
             <form action="/ctrl/article/comment.php" method="post">
                 <input type="hidden" name="article_id" value="<?= htmlspecialchars($article['id']) ?>">
                 <textarea name="content" placeholder="Votre commentaire" required></textarea>
@@ -40,12 +43,11 @@
                 <div>
                     <p><strong><?= htmlspecialchars($comment['name']) ?></strong> a commenté le <?= htmlspecialchars($comment['date']) ?></p>
                     <p><?= nl2br(htmlspecialchars($comment['textOfComment'])) ?></p>
-                    <?php if ($_SESSION['user']['id'] == $comment['idUser'] || $_SESSION['user']['idRole'] == 'admin'): ?>
+                    <?php if ($_SESSION['user']['id'] == $comment['idUser'] || $_SESSION['user']['idRole'] == '10'): ?>
                         <a href="/ctrl/article/comment.php?action=delete&id=<?= htmlspecialchars($comment['id']) ?>" onclick="return confirm('Êtes-vous sûr de vouloir supprimer ce commentaire ?');">
                             <button class="buttonDelete">Supprimer</button>
                         </a>
                         <a href="/ctrl/article/comment.php?action=edit&id=<?= htmlspecialchars($comment['id']) ?>"><button class="buttonUpdate">Modifier</button></a>
-
                     <?php endif; ?>
                 </div>
             <?php endforeach; ?>
